requirements.txt
llama_index==0.8.18
spacy==3.6.1
textbase-client>=0.1.0
nltk==3.6.3
langchain==0.0.279
